Just dumping all changes currently...
Well, I've been doing a bunch of different stuff and keep changing and reverting and running into circular dependency issues and a bunch of other stuff. So this is a mix of different stuff and very broken, as I work on fixing it, because I don't know what's what, this could take a while. I am trying to change over the very nested folder approach to a more class-based approach.