Add hotReload manifest key for plugins
You can now overwrite the user's settings if your plugin /must/ have hot reload enabled or disabled to function. It is set with an enable key boolean like:
"hotReload": {
  "enable": true
}
It is an optional key that overrides the user's hot reload setting. You may also now specify files and directories to ignore in your plugin when hot reloading with the optional ignore key, which takes a string or array argument, like:
"hotReload": {
  "enable": true,
  "ignore": [ "*.jsx", "some-folder" ]
}